requirements.txt
certifi==2019.11.28
chardet==3.0.4
Django==2.2.8
django-modeladmin-reorder==0.3.1
django-preferences==1.0.0
idna==2.8
mysqlclient==1.4.5
Pillow==6.2.1
PySocks==1.7.1
pyTelegramBotAPI==3.6.6
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
python-dotenv==0.10.3
